requirements.txt
tensorflow == 2.17.0
fastapi == 0.114.2
uvicorn == 0.30.6
numpy == 1.26.4
joblib == 1.4.2
pillow == 10.4.0

# pip install -r requirements.txt
# pip show <package-name>